Some fixes and adjustements

+++ src/main/java/com/profesorfalken/jsensors/model/sensors/Fan.java
@@ -21,9 +21,9 @@ package com.profesorfalken.jsensors.model.sensors;
  */
 public class Fan {
     public final String name;
-    public final Double value;
+    public final Integer value;
 
-    public Fan(String name, Double value) {
+    public Fan(String name, Integer value) {
         this.name = name;
         this.value = value;
     }
